+++ client/beadledom-client/src/main/java/com/cerner/beadledom/client/OffsetPaginatedListDto.java
@@ -1,4 +1,4 @@
-package com.cerner.beadledom.client.example.model;
+package com.cerner.beadledom.client;
 
 import com.cerner.beadledom.pagination.OffsetPaginatedList;
